NP-2213 Refactor contributors
